mcp/docker/pyproject.toml
[project]
name = "docker-mcp"
version = "0.1.0"
description = "Model Context Protocol Server for Docker Container Management"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
urls = { Homepage = "https://github.com/tomdwipo/agent" }
keywords = ["docker", "containers", "mcp", "devops", "automation"]
dependencies = [
    "mcp>=1.9.3",
    "docker>=7.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "rich>=13.7.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]